+++ data.tf
@@ -3,6 +3,7 @@ data "cato_accountSnapshotSite" "gcp-site" {
 }
 
 data "cato_accountSnapshotSite" "gcp-site-for-secondary" {
+  count      = var.ha ? 1 : 0
   depends_on = [time_sleep.secondary_serial_delay]
   id         = cato_socket_site.gcp-site.id
 }